requirements.txt
pyserial==3.5
numpy==2.0.2
pyaudio==0.2.14
